// apps/web/src/app/api/tasks/schemas.ts
import { z } from "zod";

export const createTaskSchema = z.object({
  title: z.string().min(1),
  description: z.string().optional(),
  status: z.enum(["TODO","DOING","DONE"]),
});

export const updateTaskSchema = z.object({
  title: z.string().optional(),
  description: z.string().optional(),
  status: z.enum(["TODO","DOING","DONE"]).optional(),
});